+++ Gemfile
@@ -3,6 +3,7 @@ source 'https://rubygems.org'
 gem 'rake'
 gem 'minitest'
 gem 'minitest-reporters'
+gem 'minitest-focus'
 gem 'capybara'
 gem 'capybara_minitest_spec'
 # gem 'selenium-webdriver'
