+++ Rakefile
@@ -8,7 +8,7 @@ namespace :load do
   desc 'Load test dovetail stitching'
   task :dovetail, [:total, :concurrency, :per] do |t, args|
     $LOAD_PATH.unshift("#{File.dirname(__FILE__)}/load")
-    $LOAD_PATH.unshift("#{File.dirname(__FILE__)}/test")
+    $LOAD_PATH.unshift("#{File.dirname(__FILE__)}/test/support")
     require 'dovetail/file_cache'
 
     args.with_defaults(total: 10, concurrency: 5, per: 1)
@@ -19,15 +19,15 @@ namespace :load do
 end
 
 Rake::TestTask.new('test') do |t|
-  t.libs << 'test'
+  t.libs << 'test/support'
   t.pattern = "test/**/*_test.rb"
 end
 Rake::TestTask.new('test:dovetail') do |t|
-  t.libs << 'test'
+  t.libs << 'test/support'
   t.pattern = "test/dovetail/*_test.rb"
 end
 Rake::TestTask.new('test:publish') do |t|
-  t.libs << 'test'
+  t.libs << 'test/support'
   t.pattern = "test/publish/*_test.rb"
 end
