+++ test/support/test_helper.rb
@@ -2,6 +2,7 @@ require 'minitest/reporters'
 require 'minitest/autorun'
 require 'minitest/spec'
 require 'minitest/pride'
+require 'minitest/focus'
 require 'excon'
 require 'json'
 require 'config'
